+++ CHANGELOG.md
@@ -12,6 +12,7 @@ vX.X.X - TBD
 * Adds a new `debug` module to the standard library which provides logging,
   assert, and unit testing functionality all in one place.
 * Adds more missing Sphere v1 API functions, further improving compatibility.
+* Renames `defScenelet()` from the Scenes module to `Scene.defineAction()`.
 * Removes the `pact` and `logger` modules from the standard library.
 * Removes the `minify` module from Cell's standard library.
 * Removes `describe()` from the Cellscript API.